Added some behavior
if the train is stopped and I have already set a speed and the LCA is 0%; if I move the LCM, the set speed will immediately become 0 so I can give traction with the LCM lever.
Second one. if the train is moving and I have already set a speed and the LCA is not 0%, I have to bring the LCA to 0% and then I can then move the LCM to accelerate the train or brake it. Obviously, as soon as I move the LCM the set speed becomes 0 and I will have to set it again if I want to use the LCA lever again.